# signrhflow-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SignRhFlow API",
        "description": "Documentacao inicial dos endpoints de colaboradores e contratos.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/contracts": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Lista contratos",
                "operationId": "listContracts",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de contratos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedContractsResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Cria contrato",
                "operationId": "storeContract",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "employee_id",
                                    "delivery_method"
                                ],
                                "properties": {
                                    "employee_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "delivery_method": {
                                        "type": "string",
                                        "enum": [
                                            "EMAIL",
                                            "WHATSAPP"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "DRAFT",
                                            "PENDING",
                                            "SIGNED",
                                            "REJECTED"
                                        ]
                                    },
                                    "file_path": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contrato criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContractResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validacao",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/contracts/{contract}": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Detalha contrato por ID",
                "operationId": "showContract",
                "parameters": [
                    {
                        "name": "contract",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contrato encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContractResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contrato nao encontrado"
                    }
                }
            }
        },
        "/api/contracts/{contract}/pdf": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Baixa PDF do contrato",
                "operationId": "downloadContractPdf",
                "parameters": [
                    {
                        "name": "contract",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arquivo PDF do contrato"
                    },
                    "404": {
                        "description": "PDF nao encontrado"
                    }
                }
            }
        },
        "/api/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Lista colaboradores",
                "operationId": "listEmployees",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de colaboradores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedEmployeesResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Cria colaborador",
                "operationId": "storeEmployee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "cpf"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Joao Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao.silva@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "(11) 99999-9999"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "52998224725"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Colaborador criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validacao",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/employees/{employee}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Detalha colaborador por ID",
                "operationId": "showEmployee",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Colaborador encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Colaborador nao encontrado"
                    }
                }
            }
        },
        "/api/webhooks/autentique": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Recebe webhook da Autentique",
                "operationId": "autentiqueWebhook",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook recebido"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "EmployeeResource": {
                "required": [
                    "id",
                    "name",
                    "email",
                    "phone",
                    "cpf"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "cpf": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ContractResource": {
                "required": [
                    "id",
                    "employee_id",
                    "status",
                    "delivery_method",
                    "file_path"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "employee_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "autentique_document_id": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "DRAFT",
                            "PENDING",
                            "SIGNED",
                            "REJECTED"
                        ]
                    },
                    "delivery_method": {
                        "type": "string",
                        "enum": [
                            "EMAIL",
                            "WHATSAPP"
                        ]
                    },
                    "file_path": {
                        "type": "string"
                    },
                    "employee": {
                        "$ref": "#/components/schemas/EmployeeResource"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "PaginatorLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginatorMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "from": {
                        "type": "integer",
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "path": {
                        "type": "string"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "to": {
                        "type": "integer",
                        "nullable": true
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "PaginatedEmployeesResponse": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EmployeeResource"
                        }
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginatorLinks"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginatorMeta"
                    }
                },
                "type": "object"
            },
            "PaginatedContractsResponse": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ContractResource"
                        }
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginatorLinks"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginatorMeta"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Employees",
            "description": "Gestao de colaboradores"
        },
        {
            "name": "Contracts",
            "description": "Gestao de contratos"
        },
        {
            "name": "Webhooks",
            "description": "Eventos de integracao externa"
        }
    ]
}